Refactoring Battle AI. Divide BattleAI on the smaller files.
