+++ AI/BattleAI/CMakeLists.txt
@@ -6,7 +6,13 @@ include_directories(${Boost_INCLUDE_DIRS} ${CMAKE_HOME_DIRECTORY} ${CMAKE_HOME_D
 set(battleAI_SRCS
 		StdInc.cpp
 		BattleAI.cpp
+		StackWithBonuses.cpp
+		EnemyInfo.cpp
+		AttackPossibility.cpp
+		PotentialTargets.cpp
 		main.cpp
+		common.cpp
+		ThreatMap.cpp
 )
 
 add_library(BattleAI SHARED ${battleAI_SRCS})
@@ -16,6 +22,6 @@ set_target_properties(BattleAI PROPERTIES ${PCH_PROPERTIES})
 cotire(BattleAI)
 
 if (NOT APPLE) # Already inside vcmiclient bundle
-    install(TARGETS BattleAI RUNTIME DESTINATION ${AI_LIB_DIR} LIBRARY DESTINATION ${AI_LIB_DIR})
+	install(TARGETS BattleAI RUNTIME DESTINATION ${AI_LIB_DIR} LIBRARY DESTINATION ${AI_LIB_DIR})
 endif()
