+++ AI/BattleAI/StdInc.h
@@ -1,5 +1,13 @@
+/*
+ * StdInc.h, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
 #pragma once
-
 #include "../../Global.h"
 
 // This header should be treated as a pre compiled header file(PCH) in the compiler building settings.
