+++ AI/BattleAI/ThreatMap.h
@@ -1,0 +1,26 @@
+/*
+ * ThreatMap.h, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
+/*
+#pragma once
+
+#include "common.h"
+#include "../../lib/BattleState.h"
+#include "CCallback.h"
+/*
+class ThreatMap
+{
+public:
+	std::array<std::vector<BattleAttackInfo>, GameConstants::BFIELD_SIZE> threatMap; // [hexNr] -> enemies able to strike
+
+	const CStack *endangered;
+	std::array<int, GameConstants::BFIELD_SIZE> sufferedDamage;
+
+	ThreatMap(const CStack *Endangered);
+};*/ // These lines may be usefull but they are't used in the code.
