+++ AI/BattleAI/main.cpp
@@ -1,5 +1,13 @@
+/*
+ * main.cpp, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
 #include "StdInc.h"
-
 #include "../../lib/AI_Base.h"
 #include "BattleAI.h"
