+++ AUTHORS
@@ -57,3 +57,6 @@ Arseniy Shestakov aka SXX,      <me@arseniyshestakov.com>
 
 Vadim Markovtsev, <gmarkhor@gmail.com>
    * resolving problems with macOS, bug fixes
+
+Michał Kalinowski, <feniks_fire@o2.pl>
+   * refactoring code
